+++ storage/fuzz/fuzz_targets/current_crash_recovery.rs
@@ -104,7 +104,7 @@ fn make_config(
             metadata_partition: format!("crash-merkle-metadata-{suffix}"),
             items_per_blob: NZU64!(merkle_items_per_blob),
             write_buffer,
-            replay_buffer: commonware_utils::NZUsize!(4096),
+            replay_buffer: NZUsize!(4096),
             strategy: Sequential,
             page_cache: page_cache.clone(),
         },
@@ -112,7 +112,7 @@ fn make_config(
             partition: format!("crash-log-{suffix}"),
             items_per_section: NZU64!(log_items_per_blob),
             write_buffer,
-            replay_buffer: commonware_utils::NZUsize!(4096),
+            replay_buffer: NZUsize!(4096),
             compression: None,
             codec_config: ((), ()),
             page_cache,
