+++ storage/fuzz/fuzz_targets/journal_crash_recovery.rs
@@ -341,7 +341,7 @@ impl FuzzJournal for FixedJournal<deterministic::Context, Item> {
                 params.page_cache_size,
             ),
             write_buffer: params.write_buffer,
-            replay_buffer: commonware_utils::NZUsize!(4096),
+            replay_buffer: NZUsize!(4096),
         }
     }
 
@@ -409,7 +409,7 @@ impl FuzzJournal for VariableJournal<deterministic::Context, Item> {
                 params.page_cache_size,
             ),
             write_buffer: params.write_buffer,
-            replay_buffer: commonware_utils::NZUsize!(4096),
+            replay_buffer: NZUsize!(4096),
         }
     }
